Simplify release script
[Card](https://rentpath.atlassian.net/browse/SRV-569)

I noticed that the release failed to build the git tag
on the jenkins slaves it happened to run on.
(Instead of `v0.0.3` the tag was just `v`.)
After investigating I see that the release script (which I copied
from one of our other libs) is doing some creative scripting
that runs clojure via a specific jar file which it turns out is missing
on some of the slaves (why the slaves differ is another matter).

Anyway there's no need to call out to clojure just to pluck out the
version from the project file like this; the same thing could be
accomplished with some good old unix commands.